dead

+++ backend/src/main/scala/com/neo/sk/carnie/paperClient/GridOnServer.scala
@@ -66,7 +66,7 @@ class GridOnServer(override val boundary: Point) extends Grid {
       case (p, f@Field(_)) => (p, f)
       case _ => (Point(-1, -1), Field((-1L).toString))
     }.filter(_._2.id != -1L).values.groupBy(_.id).map(p => (p._1, p._2.size))
-    currentRank = snakes.values.map(s => Score(s.id, s.name, s.kill, areaMap.getOrElse(s.id, 0))).toList.sortBy(_.area).reverse
+    currentRank = snakes.values.map(s => Score(s.id, s.name, s.kill, areaMap.getOrElse(s.id, 0).toShort)).toList.sortBy(_.area).reverse
 
   }
 
@@ -126,7 +126,7 @@ class GridOnServer(override val boundary: Point) extends Grid {
         if (currentRank.exists(_.id == id)) {
           val info = currentRank.filter(_.id == id).head
           (id, info.k, info.area)
-        } else (id, -1, -1)
+        } else (id, -1.toShort, -1.toShort)
       }
       roomManager ! UserDead(roomId, mode, deadSnakesInfo)
 //      log.debug(s"!!!!!!!dead snakes: ${deadSnakesInfo.map(_._1)}")
